make everything float-configurable

+++ src/color.rs
@@ -1,12 +1,12 @@
-use crate::vec::Color;
+use crate::{vec::Color, Float};
 use std::io::Write;
 
 pub fn write<W: Write>(stream: &mut W, pixel_color: Color, samples_per_pixel: u32) {
-    #![allow(clippy::cast_possible_truncation)]
+    #![allow(clippy::cast_possible_truncation, clippy::cast_precision_loss)]
     let mut pixel_color = pixel_color;
 
     // Divide the color by the number of samples.
-    let scale = 1.0 / f64::from(samples_per_pixel);
+    let scale = 1.0 / samples_per_pixel as Float;
     pixel_color *= scale;
 
     // Gamma-correct for gamma=2.
